simple-webapp-mysql/requirements.txt
Flask
mysql-connector-python
boto3
awscli